Refactored Main
All the functions were broken down into the Engine.h:
Created Window.h for method responsible for creating a window.
Created EventHandler for methods responsible for handling keyboard/mouse inputs.
IncludeHeaders.h is just a precaution header so that the order for the glfw and glad is correct.
Added shaders with the .shader extention for better usability and readability.