Updated UI.h object creation window
Added dropdown menu for choosing the object to be created
Added tabs for object manipulation before its saving
Added buttons for Creating object, Saving it to Level, Deleting it from the Level

Tabs are fulfilled by dummy buttons.
Save and Delete buttons are dummy.

All the dummies need to be added by functionality and some refactoring may take a place.